clocksource: Fix brown-bag boolean thinko in cs_watchdog_read()
The current "nretries > 1 || nretries >= max_retries" check in
cs_watchdog_read() will always evaluate to true, and thus pr_warn(), if
nretries is greater than 1.  The intent is instead to never warn on the
first try, but otherwise warn if the successful retry was the last retry.

Therefore, change that "||" to "&&".

Reported-by: Borislav Petkov <bp@alien8.de>
Fixes: db3a34e17433 ("clocksource: Retry clock read if long delays detected")
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>