rcu: Make __sync_rcu_exp_select_node_cpus() verify same grace period
One suspicious property of the __sync_rcu_exp_select_node_cpus()
function is that it drops the rcu_node structure’s lock between
detecting a set of offline CPUs and reporting quiescent states on
their behalf.  Obviously, a preemption, interrupt, or whatever
can occur during that interval, which might allow that CPU to come
back online.  If it does, the rcutree_online_cpu() function will
invoke sync_sched_exp_online_cleanup(), which can (among other equally
deleterious options) invoke smp_call_function_single() to force an IPI
invoking rcu_exp_handler() on the incoming CPU, which could well in turn
report a quiescent state for the incoming CPU.

Which is fine.  Except that the current expedited RCU grace
period might end and another one might start, both before the
__sync_rcu_exp_select_node_cpus() function gets around to reporting
the quiescent state on the incoming CPU’s behalf.  In this case,
a quiescent state from the old (and now ended) grace period will be
incorrectly reported against the new grace period, possibly resulting
in a too-short grace period.

This commit therefore passes the current grace period sequence
number down through a cascade of functions in order to permit the
__sync_rcu_exp_select_node_cpus() function to check whether that grace
period is still in effect.  If it is not, this function takes an early
exit, thus avoiding the too-short grace periods.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>
Cc: Frederic Weisbecker <frederic@kernel.org>