rcutorture: Collect CPU ID for decoration after changing reader type
Currently, rcutorture_one_extend() collects the CPU ID before making
any change to the type of RCU reader.  This can be confusing because the
properties of the code from which the CPU ID is collected is not that of
the reader segment that this same CPU ID is listed with.  This commit
therefore causes rcutorture_one_extend() to collect the CPU ID just
after the type of RCU reader has been changed.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>