ratelimit: Create functions to handle the ->missed field
The ratelimit_state structure's ->missed field is accessed open-coded by
a number of ratelimit users.  This works, but is a bit messy and makes
it more difficult to make changes to this field.

Therefore, provide a ratelimit_state_get_miss() function that reads
out the ->missed field of the specified ratelimit_state structure, and
also a ratelimit_state_reset_miss() function that reads out that field,
but that also resets its value to zero.  These functions will replace
client-code open-coded uses of ->miss.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>
Cc: Andrew Morton <akpm@linux-foundation.org>
Cc: Kuniyuki Iwashima <kuniyu@amazon.com>
Cc: Mateusz Guzik <mjguzik@gmail.com>