rcutorture: Check for no up reader at task level
The design of testing of up/down readers such as srcu_down_read()
and srcu_up_read() assumes that these are tested only by the
rcu_torture_updown() kthread, and never by the rcu_torture_reader()
kthread.  Because we all know which road is paved with good intentions,
this commit adds WARN_ON_ONCE() to verify that the ->up_read() really
is executing in either a softirq or an interrupt handler.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>