squash! rcutorture: Complain if an ->up_read() is delayed more than 10 seconds
This to-be-squashed commit avoids false-positive complaint of excessive
->up_read() delays.  These false positives could arise by task-level
delays in the rcu_torture_updown() and rcu_torture_updown_one() functions,
or due to delays in softirq scheduling of the HRTIMER_MODE_SOFT
rcu_torture_updown_hrt() handler.  Avoid the task-level delays by
sampling the time after queuing the hrtimer on the one hand and by
sampling the time before reading the ->rtorsu_inuse flag on the other.
Avoid the softirq scheduling delays through use of HRTIMER_MODE_HARD.

[ paulmck: Apply kernel test robot feedback. ]

Tested-by: kernel test robot <oliver.sang@intel.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>