torture: Make torture.sh KCSAN runs set CONFIG_RCU_TORTURE_TEST_CHK_RDR_STATE=y
No rcutorture testing sets CONFIG_RCU_TORTURE_TEST_CHK_RDR_STATE=y,
mostly because this was intended to be a deep debugging Kconfig option.
And it does have high read-side overhead.  But on the other hand, it
would have found an rcutorture bug in the testing of srcu_up_read()
and if it is not tested regularly, it won't be working when we need it.

This commit therefore enables this Kconfig option in all of the torture.sh
script's KCSAN runs.

Reported-by: kernel test robot <oliver.sang@intel.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>