rcutorture: Build with CONFIG_RCU_TORTURE_TEST_CHK_RDR_STATE=y for KCSAN
The RCU_TORTURE_TEST_CHK_RDR_STATE Kconfig option is used for
low-level debugging of rcutorture's generation of overlapping and
nested RCU readers.  It incurs significant overhead, and is thus not
to be used lightly.  But if it is not tested regularly, it won't
be there when it is needed.  On the other hand, KCSAN's overhead
is also non-trivial.  This commit therefore builds KCSAN kernels
CONFIG_RCU_TORTURE_TEST_CHK_RDR_STATE=y, but only for the --d0-rcutorture
case.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>