squash! tracing: Guard __DECLARE_TRACE() use of __DO_TRACE_CALL() with SRCU-fast
@@@ Adjust commit message:

However, Yonghong Song points out that BPF expects non-sleepable BPF
programs to remain on the same CPU, which means that migration must
be disabled whenever preemption remains enabled.  In addition, non-RT
kernels have performance expectations on BPF that would be violated
by allowing the BPF programs to be preempted.

Therefore, continue to disable preemption in non-RT kernels, and protect
the BPF program with both SRCU and migration disabling for RT kernels,
and even then only if preemption is not already disabled.

[ paulmck: Apply Yonghong Song feedback. ]

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>