srcu: Make DEFINE_SRCU_FAST() available to modules
This commit makes DEFINE_SRCU_FAST() available to modules.  The trick
is that DEFINE_SRCU_FAST() does not allocate the per-CPU data structure
when used within a module (by design), which means that it is not possible
to compile-time-initialize any of the fields in this structure, including
in particular the ->srcu_reader_flavor field.

This commit therefore creates an ->srcu_reader_flavor field in the
srcu_struct structure, adds arguments to the DEFINE_SRCU()-related
macros to initialize this new field, and extending the checks in the
__srcu_check_read_flavor() function to include this new field.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>
Cc: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Cc: Steven Rostedt <rostedt@goodmis.org>
Cc: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Cc: <bpf@vger.kernel.org>

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>