srcu: Optimize SRCU-fast for arm64
Some arm64 platforms have slow per-CPU atomic operations, for example, the
Neoverse V2.  This commit therefore moves SRCU-fast from per-CPU atomic
operations to preemption-disabled non-atomic READ_ONCE()/WRITE_ONCE()
operations.  This works because SRCU-fast does not provide up/down
read-side primitives, which means that if srcu_read_unlock_fast()
is in (say) an NMI handler, then the matching srcu_read_lock_fast()
must also have been in that same NMI handler.  This means that
srcu_read_unlock_fast() can check for being in a softirq, interrupt,
or NMI handler, and if it is, it can decrement the lock count instead
of incrementing the unlock count.  This works because it is not possible
to migrate from one CPU to another while executing in one of these types
of handlers.

TBD performance data once debugged.  Expected to reduce from about 100ns
to about 3ns on ARM, and to remain about the same on x86.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: Will Deacon <will@kernel.org>
Cc: Mark Rutland <mark.rutland@arm.com>
Cc: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Cc: Steven Rostedt <rostedt@goodmis.org>
Cc: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Cc: <linux-arm-kernel@lists.infradead.org>
Cc: <bpf@vger.kernel.org>